+++ tests/Silex/Tests/Provider/TranslationServiceProviderTest.php
@@ -50,6 +50,7 @@ class TranslationServiceProviderTest extends \PHPUnit_Framework_TestCase
                 )
             )
         );
+
         return $app;
     }